+++ lib/src/cli/option.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2021 - 2023 Buijs Software
+// Copyright (c) 2021 - 2024 Buijs Software
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -23,52 +23,73 @@ import "dart:io";
 import "../common/common.dart";
 import "context.dart";
 
+/// An option containing a String value which should be declared by the user.
+///
+/// For this option no defaults can be set.
 class RequiredUserInputString extends RequiredUserInput<String> {
+  /// Construct a new [RequiredUserInputString] instance.
   const RequiredUserInputString(this.description);
 
+  @override
   final String description;
 
+  @override
   String convertOrThrow(String value) => value;
 }
 
+/// An option containing a bool value which has a default value.
 class UserInputBooleanOrDefault extends UserInputOrDefault<bool> {
+  /// Construct a new [RequiredUserInputString] instance.
   const UserInputBooleanOrDefault(this.description, super.defaultValue);
 
+  @override
   final String description;
 
-  bool convertOrThrow(String value) {
-    switch (value.trim().toUpperCase()) {
-      case "TRUE":
-        return true;
-      case "FALSE":
-        return false;
-      default:
-        throw InputException("expected a bool value but received: $value");
-    }
-  }
+  static const _map = {"TRUE": true, "FALSE": false};
+
+  @override
+  bool convertOrThrow(String value) =>
+      _map[value.trim().toUpperCase()] ??
+      (throw InputException("expected a bool value but received: $value"));
 }
 
+/// An option containing a value which should be declared by the user.
+///
+/// For this option no defaults can be set.
 abstract class RequiredUserInput<T> extends Input<T> {
+  /// Construct a new [RequiredUserInput] instance.
   const RequiredUserInput();
 }
 
+/// An option containing a [T] value which has a default value.
 abstract class UserInputOrDefault<T> extends Input<T> {
+  /// Construct a new [UserInputOrDefault] instance.
   const UserInputOrDefault(this.defaultValue);
 
+  /// Default value [T] which is returned when user did not specify the option.
   final T defaultValue;
 
+  /// Get the [T] defaultValue as String.
+  ///
+  /// Can be overwritten by extending class to apply custom serialization.
   String get defaultValueToString => defaultValue.toString();
 }
 
+/// An option containing a [T] value.
 abstract class Input<T> {
+  /// Construct a new [Input] instance.
   const Input();
 
-  /// Throws [InputException].
+  /// Convert a String value to type [T] or throw [InputException].
   T convertOrThrow(String value);
 
+  /// A description for this option.
   String get description;
 }
 
+/// Exception indicating an error in the user input.
+///
+/// This exception occurs when the user input can not be mapped to [Input].
 class InputException implements Exception {
   /// Create instance of [InputException] with a message [cause].
   InputException(this.cause);
@@ -98,7 +119,11 @@ class RootDirectoryInput
   String get description => "the klutter project root directory";
 }
 
+/// Option specifying the flutter version to use.
+///
+/// Containing a [VerifiedFlutterVersion] value which defaults to flutter 3.10.6.
 class FlutterVersionOption extends UserInputOrDefault<VerifiedFlutterVersion> {
+  /// Construct a new [FlutterVersionOption] instance with default value of 3.10.6.
   FlutterVersionOption() : super("3.10.6".verifyFlutterVersion!);
 
   @override
@@ -115,20 +140,30 @@ class FlutterVersionOption extends UserInputOrDefault<VerifiedFlutterVersion> {
           "invalid flutter version (supported versions are: ${supportedFlutterVersions.keys}): $value"));
 }
 
+/// Option to allow overwriting existing entities which defaults to false.
 class OverwriteOption extends UserInputBooleanOrDefault {
+  /// Construct a new [OverwriteOption] instance.
   OverwriteOption()
       : super("overwrite existing distribution when found", false);
 }
 
+/// Option to skip download actions which defaults to false.
+///
+/// This options is used for testing download actions.
 class DryRunOption extends UserInputBooleanOrDefault {
+  /// Construct a new [DryRunOption] instance.
   DryRunOption() : super("skip downloading of libraries", false);
 }
 
+/// Option to specify a klutter library name which has no default.
 class LibraryName extends RequiredUserInputString {
+  /// Construct a new [LibraryName] instance.
   const LibraryName() : super("name of the library to add");
 }
 
+/// Option to specify squint_json version which defaults to [squintPubVersion].
 class SquintPubVersion extends UserInputOrDefault<String> {
+  /// Construct a new [SquintPubVersion] instance with default value [squintPubVersion].
   const SquintPubVersion() : super(squintPubVersion);
 
   @override
@@ -138,7 +173,9 @@ class SquintPubVersion extends UserInputOrDefault<String> {
   String convertOrThrow(String value) => value;
 }
 
+/// Option to specify klutter (dart) version which defaults to [klutterPubVersion].
 class KlutterPubVersion extends UserInputOrDefault<String> {
+  /// Construct a new [KlutterPubVersion] instance with default value [klutterPubVersion].
   const KlutterPubVersion() : super(klutterPubVersion);
 
   @override
@@ -148,7 +185,9 @@ class KlutterPubVersion extends UserInputOrDefault<String> {
   String convertOrThrow(String value) => value;
 }
 
+/// Option to specify klutter_ui (dart) version which defaults to [klutterUIPubVersion].
 class KlutteruiPubVersion extends UserInputOrDefault<String> {
+  /// Construct a new [KlutteruiPubVersion] instance with default value [klutterUIPubVersion].
   const KlutteruiPubVersion() : super(klutterUIPubVersion);
 
   @override
@@ -158,7 +197,9 @@ class KlutteruiPubVersion extends UserInputOrDefault<String> {
   String convertOrThrow(String value) => value;
 }
 
+/// Option to specify klutter (gradle) version which defaults to [klutterGradleVersion].
 class KlutterGradleVersionOption extends UserInputOrDefault<String> {
+  /// Construct a new [KlutterGradleVersionOption] instance with default value [klutterGradleVersion].
   const KlutterGradleVersionOption() : super(klutterGradleVersion);
 
   @override
@@ -170,7 +211,9 @@ class KlutterGradleVersionOption extends UserInputOrDefault<String> {
       (throw InputException("not a valid bom version: $value"));
 }
 
+/// Option to specify plugin_name which defaults to my_plugin.
 class PluginNameOption extends UserInputOrDefault<String> {
+  /// Construct a new [PluginNameOption] instance with default value my_plugin.
   const PluginNameOption() : super("my_plugin");
 
   @override
@@ -187,7 +230,9 @@ class PluginNameOption extends UserInputOrDefault<String> {
   }
 }
 
+/// Option to specify group_name which defaults to dev.buijs.klutter.example.
 class GroupNameOption extends UserInputOrDefault<String> {
+  /// Construct a new [GroupNameOption] instance with default value dev.buijs.klutter.example.
   GroupNameOption() : super("dev.buijs.klutter.example");
 
   @override