+++ test/src/cli/task_project_init_test.dart
@@ -1,0 +1,68 @@
+// Copyright (c) 2021 - 2024 Buijs Software
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+import "dart:io";
+
+import "package:klutter/klutter.dart";
+import "package:klutter/src/cli/context.dart";
+import "package:test/test.dart";
+
+void main() {
+
+  test("Verify init in consumer project skips producer init", () async {
+    final pathToRoot =
+        Directory("${Directory.systemTemp.absolute.path}/build_test".normalize)
+          ..createSync();
+
+    final project = await createFlutterProjectOrThrow(
+      executor: Executor(),
+      pathToFlutter: "flutter",
+      pathToRoot: pathToRoot.absolutePath,
+      name: "my_project",
+      group: "my.org",
+    );
+
+    final task = ProjectInit();
+    final context = Context(workingDirectory: project.resolveFolder("example"),
+        taskName: TaskName.init,
+        taskOptions: {});
+    final result = await task.execute(context);
+    expect(result.isOk, true);
+
+    final mainDartFile =  File("${project.absolutePath}/example/lib/main.dart"
+        .normalize);
+    expect(mainDartFile.existsSync(),
+        true,
+        reason: "example/lib/main.dart file should exist");
+
+    expect(mainDartFile
+        .readAsStringSync()
+        .contains('String _greeting = "There shall be no greeting for now!";'),
+        true,
+        reason: "main.dart content is overwritten");
+
+    final registry =
+    File("${project.absolutePath}/example/.klutter-plugins".normalize);
+
+    expect(registry.existsSync(), true,
+        reason: "klutter-plugins file should be created");
+
+  });
+}